requirements.txt
tiktoken==0.4.0
openai==0.27.8
python-dotenv==1.0.0
chromadb==0.3.26
langchain==0.0.220
unstructured==0.7.12
libmagic==1.0
SpeechRecognition==3.10.0
PyAudio==0.2.13
pyttsx==2.90
pyobjc==9.2
